GenomicDiversity.jl

For an overview of this package with an example of its use, see here.

Below are descriptions of all functions in the package.

Functions

GenomicDiversity.GenoData — Type

GenoData(indInfo::DataFrame, genotypes::Matrix, positions::DataFrame)

Constructs a GenoData struct, which stores metadata of individuals, a genotype matrix of individuals and SNPS, and genomic position info.

​# Arguments / keywords

indInfo: A DataFrame containing metadata.

genotypes: A Matrix containing genotypes for individuals (rows) and loci (columns).

positions: A 2-column DataFrame containing genomic location info of SNPs.

Notes

You can access each component of this data structure by appending ".indInfo" or ".genotypes" or "positions" to the name of the GenoData object.

GenomicDiversity.chooseChrRegion — Method

chooseChrRegion(pos, chr; positionMin=1, positionMax=NaN)

Designate a specific region of a chromosome for subsequent analysis.

​# Arguments

pos: A matrix containing genomic location of each locus; must have a chrom column (name of scaffold) and a position column (numerical position on scaffold).

chr: The name of the chosen scaffold.

positionMin: Optional (default 1); the starting location of the chosen region.

positionMax: Optional (default the end of the scaffold); the ending location of the chosen region.

Notes

Returns a tuple containing:

the scaffold name

the minimum position

the maximum position

a string describing the region (e.g. chr Z: 20000 to 1000000)

This function does not actually filter the genotype matrix–that can be done by providing the output of this to functions such as plotGenotypeByIndividual().

GenomicDiversity.chooseChrRegion — Method

chooseChrRegion(myGenoData::GenoData, chr; positionMin=1, positionMax=NaN)

Call chooseChrRegion using a GenoData object.

GenomicDiversity.filterInds — Method

filterInds(myGenoData::GenoData, selection; direction = "out")

Filter the individuals indicated by selection, and store result in a new GenoData object. The default is to filter individuals out (remove them); set direction = "in" to filter them in (and remove others).

​# Arguments

myGenoData: A GenoData object.

selection: A vector of Boolean values, or a BitVector.

direction: (optional) The direction to filter (default is "out")

Notes

Returns a GenoData object, containing:

indInfo: A DataFrame containing metadata.

genotypes: A Matrix containing genotypes for individuals (rows) and loci (columns).

positions: A 2-column DataFrame containing genomic location info of SNPs.

GenomicDiversity.filterLoci — Method

filterLoci(myGenoData::GenoData, selection; direction = "out")

Filter the loci indicated by selection, and store result in a new GenoData object. The default is to filter loci out (remove them); set direction = "in" to filter them in (and remove others).

​# Arguments

myGenoData: A GenoData object.

selection: A vector of Boolean values, or a BitVector.

direction: (optional) The direction to filter (default is "out")

Notes

Returns a GenoData object, containing:

indInfo: A DataFrame containing metadata.

genotypes: A Matrix containing genotypes for individuals (rows) and loci (columns).

positions: A 2-column DataFrame containing genomic location info of loci.

GenomicDiversity.filterNamedInds — Method

filterNamedInds(myGenoData::GenoData, inds; direction = "out")

Filter out the individuals indicated by selection, and store result in a new GenoData object.

​# Arguments

myGenoData: A GenoData object.

inds: A collection of names (strings) to filter out.

direction: (optional) The direction to filter (default is "out")

Notes

Returns a GenoData object, containing:

indInfo: A DataFrame containing metadata.

genotypes: A Matrix containing genotypes for individuals (rows) and loci (columns).

positions: A 2-column DataFrame containing genomic location info of loci.

GenomicDiversity.getDxy — Method

getDxy(freqs, groupsToCalc)

Calculate per-locus Dxy (pi_between) for each pair of populations.

Arguments

freqs: Matrix containing alternate allele frequencies for each group (row) and SNP (column).

groupsToCalc: Vector containing the names of the groups corresponding to each row of freqs.

Notes

Returns a tuple containing:

Matrix of Dxy values for each comparison (row) and locus (column).

Vector of comparison names.

GenomicDiversity.getFreqsAndSampleSizes — Method

getFreqsAndSampleSizes(genotypes, indGroup, groupsToCalc)

Calculate allele frequencies and sample sizes for each group and SNP.

​# Arguments

genotypes: The genotype matrix, where rows are individuals and columns are loci, with genotype codes 0,1,2 meaning homozygous reference, heterozygote, homozygous alternate, and missing genotypes can be either -1 or missing.

indGroup: A vector providing the group name each individual belongs to.

groupsToCalc: A list of group names to include in calculations.

Notes

Returns a tuple containing 1) a matrix of frequencies, and 2) a matrix of samples sizes (in both, rows are groups and columns are loci).

GenomicDiversity.getFreqsAndSampleSizes — Method

getFreqsAndSampleSizes(myGenoData::GenoData, groupsToCalc)

Call getFreqsAndSampleSizes() using a GenoData object as input.

GenomicDiversity.getFst — Method

getFst(freqs, sampleSizes, groupsToCalc; among=false)

Calculate Fst for each locus, between pairs of groups and (optionally) among all groups.

Fst is calculated according to Weir & Cockerham (1984), including their correction for sample size and number of groups.

​# Arguments

freqs: Matrix containing alternate allele frequencies for each group (row) and SNP (column).

sampleSizes: Matrix containing sample sizes for each group (row) and SNP (column).

groupsToCalc: The list of group names.

among: Optional argument–set to "true" to also calculate Fst among all groups.

Notes

Returns a tuple containing:

Matrix of Fst values for each comparison (row) and locus (column).

Matrix of numerator values from the Fst calculation for each comparison (row) and locus (column).

Matrix of denominator values from the Fst calculation for each comparison (row) and locus (column).

Vector of comparison names.

GenomicDiversity.getPairwiseNames — Method

getPairwiseNames(groupsToCalc)::Vector{String}

Using a list of names, return a list of paired names in format "name1_name2".

​# Arguments

groupsToCalc: The list of names.

Notes

Returns a vector of paired names.

GenomicDiversity.getRegionDxy — Method

getRegionDxy(siteDxyMatrix)

Given a matrix representing per-population-pair (rows) per-site (columns) Dxy values (within-group nucleotide diversity), produce a vector containing the mean Dxy for each population pair.

GenomicDiversity.getRegionPi — Method

getRegionPi(sitePiMatrix)

Given a matrix representing per-group (rows) per-site (columns) Pi values (within-group nucleotide diversity), produce a vector containing the mean pi for each group.

GenomicDiversity.getRollingMean — Method

getRollingMean(inputVector, windowSize)

Calculate non-overlapping windowed arithmetic average (i.e., means) from a vector of numeric values.

​# Arguments

inputVector: A vector of numeric values (can include NaN).

windowSize: An integer indicating the window number of values to include in each mean.

Notes

Returns a vector of windowed means.

In the calculations, NaN values are ignored.

If all values in a window are NaN, then the mean is NaN.

GenomicDiversity.getSitePi — Method

getSitePi(freqs, sampleSizes)

Calculate non-biased nucleotide diversity (pi) at each site for each population, with correction for sample size, using allele freqs and sample size as input.

​# Arguments

freqs: Matrix containing alternate allele frequencies for each group (row) and SNP (column).

sampleSizes: Matrix containing sample sizes for each group (row) and SNP (column).

GenomicDiversity.getWindowBoundaries — Method

getWindowBoundaries(lociVector, windowSize)

Determine the start and end scaffold positions for each window along a single scaffold.

​# Arguments

lociVector: Vector of included locus positions along a scaffold.

windowSize: Size of each window (in number of loci).

Notes

returns a Matrix, where rows represent windows and columns are starting location and ending location of each window along the scaffold.

GenomicDiversity.getWindowedFst — Method

getWindowedFst(FstNumerator, FstDenominator, pos, windowSize)

Calculate windowed Fst along chromosomal regions, uses output from each site's Fst (produced by the getFst() function).

This is calculated according to Weir&Cockerham1984 (with sample size and pop number correction), calculated as windowed numerator over windowed denominator, in whole windows starting on left side of the region.

​# Arguments

FstNumerator: Produced by the getFst() function, this is a vector of the Fst numerator values for each locus.

FstDenominator: Produced by the getFst() function, this is a vector of the Fst denominator values for each locus.

pos: Matrix providing genomic location of each locus; there must be a position column.

windowSize: An integer indicating the window number of values to include in each mean.

Notes

Returns a tuple of two vectors: rolling mean position, and rolling mean Fst.

The calculations ignore NaN values (which would be in both numerator and denominator).

Returns NaN for a window if there are no real values in the window.

GenomicDiversity.getWindowedIndHet — Method

getWindowedIndHet(genotypes, pos, windowSize)

Calculate windowed heterozygosity for each individual, across a genomic region.

​# Arguments

genotypes: The genotype matrix, where rows are individuals and columns are loci, with genotype codes 0,1,2 meaning homozygous reference, heterozygote, homozygous alternate, and missing genotypes can be either -1 or missing.

pos: A matrix containing genomic location of each locus; must have a position column (numerical position on scaffold).

windowSize: An integer indicating the number of loci to include in each mean.

Notes

Returns a tuple containing 1) a vector of windowed mean positions, and 2) a matrix of windowed heterozgyosity values (rows are individuals and columns are loci).

GenomicDiversity.getWindowedViSHet — Method

getWindowedViSHet(windowedIndHet_standardized)

Calculate the windowed Variance in Standardized Heterozygosity (ViSHet).

​# Arguments

windowedIndHet_standardized: The matrix of standardized individual windowed heterozygosity values (rows are individuals; columns represent windowed genomic regions)

Notes

Returns a vector of variances in standardized individual heterozygosities (each value corresponds to a different windowed genomic region).

GenomicDiversity.greet_GenomicDiversity — Method

greet_GenomicDiversity()

A simple function for testing that the module has loaded.

GenomicDiversity.imputeMissingGenoData — Method

imputeMissingGenotypes(myGenoData::GenoData; method = "SVD")

When applied to a GenoData object, imputes the data in the genotypes object, and returns a new GenoData object.

GenomicDiversity.imputeMissingGenoData — Method

imputeMissingGenotypes(genotypeMatrix::Matrix; method = "SVD")

Imputes missing values in the genotype matrix.

​# Arguments

genotypeMatrix: Matrix of genotypes, where missing data are indicated with either -1 or missing.

method: The method used for imputing. Presently "SVD" is the only option (and default).

Notes

Returns the genotype matrix with filled-in missing values, of type Float32.

GenomicDiversity.limitIndsToPlot — Method

limitIndsToPlot(plotGroups, numIndsToPlot, genoData, indMetadata)

Select a subset of individuals, based on numbers within each group, for subsequent analysis.

​# Arguments

plotGroups: Vector of names of groups to include.

numIndsToPlot: Vector of maximum number of individuals to include from each group (order as in plotGroups).

genoData: Matrix of genotypes (rows are individuals, columns are loci).

indMetadata: Matrix containing metadata for individuals (make sure there is an Fst_group column).

sortByMissing: Set to true to choose individuals with the least missing data among these loci.

Notes

Returns a tuple containing:

Matrix of genotypes for included individuals.

Matrix of metadata for included individuals.

GenomicDiversity.loadGenoData — Method

loadGenoData(metadataFile, individualsFile, genotypesFile, positionsFile)

Imports data from four files and builds a GenoData object.

​# Arguments

metadataFile: The path and name of the metadata file.

individualsFile: The path and name of the file providing names of the individuals (corresponding to rows of the genotype matrix).

genotypesFile: The path and name of the file containing the matrix of genotypes.

positionsFile: The path and name of the file containing the genomic position info of the loci corresponding to columns of the genotype matrix.

Notes

Returns a GenoData object, containing:

indInfo: A DataFrame containing metadata.

genotypes: A Matrix containing genotypes for individuals (rows) and loci (columns).

positions: A 2-column DataFrame containing genomic location info of SNPs.

GenomicDiversity.plotGenomeFst — Method

plotGenomeFst(scaffolds_to_plot, 
                windowed_Fst_all,
                pairwiseNamesFst,
                windowed_pos_all,
                groupsToPlotFst,
                groupColorsFst;
                lineTransparency = 0.8,
                fillTransparency = 0.2,
                figureSize = (1200, 1200))

Plot windowed Fst along chromosomal regions, for multiple population comparisons.

​# Arguments

scaffolds_to_plot: a vector of the scaffold names to include in the plot.

windowed_Fst_all: a matrix of Fst values (rows representing different population comparisons; columns different genomic regions).

pairwiseNamesFst: a vector of names indicating the pairs of populations in rows of windowed_Fst_all.

windowed_pos_all: dataframe containing genomic locations of windows (must have columns chrom and mean_position).

groupsToPlotFst: a vector of strings indicating the population comparisons to include.

groupColorsFst: the colors to use to plot the comparisons in groupsToPlotFst.

lineTransparency (optional): transparency of the lines.

fillTransparency (optional): transparency of the fill color.

figureSize (optional): size of the figure (height, width).

Notes

Returns the figure.

GenomicDiversity.plotGenomeViSHet — Method

plotGenomeViSHet(scaffolds_to_plot, 
                windowed_ViSHet_all,
                windowed_pos_all;
                fillColor = "purple",
                lineTransparency = 0.8,
                fillTransparency = 0.2,
                figureSize = (1200, 1200),
                plotRegions = false,
                regionsToPlot = [],
                regionColor = "red")

Plot windowed Variance in Standardized Heterozygosity (ViSHet) along multiple scaffolds.

​# Arguments

scaffolds_to_plot: a vector of the scaffold names to include in the plot.

windowed_ViSHet_all: a vector of ViSHet values for windows of the genome.

windowed_pos_all: dataframe containing genomic locations of windows (must have columns chrom and mean_position).

fillColor (optional): the color for the lines and fill (default "purple").

lineTransparency (optional): transparency of the lines.

fillTransparency (optional): transparency of the fill color.

figureSize (optional): size of the figure (height, width).

plotRegions (optional): choose true to highlight regions.

regionsToPlot (optional): a DataFrame with three columns: regionChrom containing names of scaffolds, regionStart and regionEnd giving start and end of regions to highlight.

regionColor (optional): the color for the highlighted regions.

Notes

Returns the figure.

GenomicDiversity.plotGenotypeByIndividual — Method

plotGenotypeByIndividual(regionInfo, pos, genoData, 
                            indMetadata, freqs,
                            plotGroups, plotGroupColors;
                            missingFractionAllowed = 1.0,
                            colorAllelesByGroup = true, group1 = plotGroups[1],
                            indFontSize=10, figureSize=(1200, 1200),
                            show_SNP_density = true, densityPlotColor = :steelblue1,
                            plotTitle = nothing, titleFontSize = 20,
                            indColorLeftProvided = false,
                            indColorRightProvided = false)

Construct a genotype-by-individual plot, with option to filter out SNPs with too much missing data.

Under the default setting, alleles are colored (dark purple vs. light purple) according to whichever allele is designated as group1.

​# Arguments

regionInfo: Information regarding the scaffold and region of focus; a tuple provided by chooseChrRegion().

pos: Matrix providing genomic location of each locus; there must be a chrom column and a position column.

genoData: Matrix containing genotype data (individuals in rows, loci in columns).

indMetadata: Matrix of metadata for individuals; must contain Fst_group and plot_order columns.

freqs: Matrix of alternate allele frequencies for each group (row) and locus (column).

plotGroups: Vector of group names to include in plot.

plotGroupColors: Vector of plotting colors corresponding to the groups.

missingFractionAllowed: Optional; The maximum missing genotype fraction for a locus to be included (default is 1.0, so no filtering).

colorAllelesByGroup: Optional; set to false to color alleles according to reference and alternate.

group1: Optional (default is plotGroups[1]); when colorAllelesByGroup is true, this is the group that determine which allele is dark purple.

indFontSize: Optional; the font size of the individual ID labels.

figureSize: Optional; the size of the figure; default is (1200, 1200).

show_SNP_density: Optional; default is true to show a density plot.

densityPlotColor: Optional; default is :steelblue1

plotTitle: Optional; default will make a title. For no title, set to "".

titleFontSize; Optional; default is 20.

indColorLeftProvided: Optional; Default is false. Set to true if there is a column labeled indColorLeft in the metadata providing color of each individual for plotting on left side.

indColorRightProvided: Optional; same as above but for right side (requires indColorRight column in metadata).

Notes

Returns a tuple containing:

the figure

the plotted genotypes

the numerical positions (in the chosen scaffold) of the plotted loci

the sorted metadata matrix for the plotted individuals

GenomicDiversity.plotGenotypeByIndividualWithFst — Method

plotGenotypeByIndividualWithFst(groupsToCompare, Fst_cutoff, missingFractionAllowed,
                        regionInfo, pos, Fst, pairwiseNamesFst,
                        genoData, indMetadata, freqs, plotGroups, plotGroupColors;
                        colorAllelesByGroup = true, group1 = plotGroups[1],
                        indFontSize=10, figureSize=(1200, 1200),
                        show_SNP_density = true, densityPlotColor = :steelblue1,
                        plotTitle = nothing, titleFontSize = 20,
                        highlightRegionStarts = [],
                        highlightRegionEnds = [],
                        highlightRegionColor = "red")

Construct a genotype-by-individual plot, including only loci that pass thresholds for Fst and missing data.

Under the default setting, alleles are colored (dark purple vs. light purple) according to whichever allele is designated as group1.

​# Arguments

groupsToCompare: The two groups (in format name1_name2) for filtering loci based on Fst. This can be a single string, or a vector of strings, each with a population comparison to include if Fst is above Fst_cutoff.

Fst_cutoff: The minimum Fst for a locus to be included in the plot.

missingFractionAllowed: The maximum missing genotype fraction for a locus to be included.

regionInfo: Information regarding the scaffold and region of focus; a tuple provided by chooseChrRegion().

pos: Matrix providing genomic location of each locus; there must be a chrom column and a position column.

Fst: Matrix of Fst values; can be produced by getFst().

pairwiseNamesFst: Vector of pairwise names corresponding to rows in the Fst matrix.

genoData: Matrix containing genotype data (individuals in rows, loci in columns).

indMetadata: Matrix of metadata for individuals; must contain Fst_group and plot_order columns.

freqs: Matrix of alternate allele frequencies for each group (row) and locus (column).

plotGroups: Vector of group names to include in plot.

plotGroupColors: Vector of plotting colors corresponding to the groups.

colorAllelesByGroup: Optional; set to false to color alleles according to reference and alternate.

group1: Optional (default is plotGroups[1]); when colorAllelesByGroup is true, this is the group that determine which allele is dark purple.

indFontSize: Optional; the font size of the individual ID labels.

figureSize: Optional; the size of the figure; default is (1200, 1200).

show_SNP_density: Optional; default is true to show a density plot.

densityPlotColor: Optional; default is :steelblue1

plotTitle: Optional; default will make a title. For no title, set to "".

titleFontSize; Optional; default is 20.

highlightRegionStarts: Optional; the left locations of regions to highlight on the scaffold.

highlightRegionEnds: Optional; the right locations of regions to highlight on the scaffold.

highlightRegionColor: Optional; default red; the color of the highlight bar along the scaffold.

Notes

Returns a tuple containing:

the figure

the plotted genotypes

the numerical positions (in the chosen scaffold) of the plotted loci

the sorted metadata matrix for the plotted individuals

GenomicDiversity.plotGenotypeByIndividualWithFst — Method

plotGenotypeByIndividualWithFst(myGenoData::GenoData, groupsToCompare, 
                        Fst_cutoff, missingFractionAllowed,
                        regionInfo, Fst, pairwiseNamesFst,
                        freqs, plotGroups, plotGroupColors;
                        colorAllelesByGroup = true, group1 = plotGroups[1],
                        indFontSize=10, figureSize=(1200, 1200),
                        show_SNP_density = true, densityPlotColor = :steelblue1,
                        plotTitle = nothing, titleFontSize = 20,
                        highlightRegionStarts = [],
                        highlightRegionEnds = [],
                        highlightRegionColor = "red")

Call the plotGenotypeByIndividualWithFst() function using a GenoData object.

GenomicDiversity.plotPCA — Method

plotPCA(genotypes, ind_with_metadata, groups_to_plot_PCA, group_colors_PCA; sampleSet = "", regionText="", flip1 = false, flip2 = false)

Do a principal components analysis based on genotypes of individuals (colored by group), and display PC1 vs. PC2.

​# Arguments

genotypes: Genotype matrix (individuals in rows, loci in columns, with no missing data–can impute prior to this, to ensure no missing).

indMetadata: Matrix containing metadata for individuals (make sure there is an Fst_group` column).

groups_to_plot_PCA: Vector of names of groups to include.

group_colors_PCA: Vector listing plotting color for each each group.

Optional arguments:

sampleSet: Name of the sample set to appear in the plot title.

regionText: Name of the genomic region to appear in the plot title.

flip1: Set to true if wanting to flip PC1 (i.e., multiply by -1).

flip2: Same but for PC2.

showPlot: Set to false if not wanting the function to draw display.

autolimitaspect_setting: For unconstrained axes, set to nothing (default is 1 for 1:1 axes).

lineOpacity: Opacity of the line color of symbols.

fillOpacity: Opacity of the fill color of symbols.

symbolSize: Size of symbols.

plotTitle: Specify if a specific title is wanted; otherwise will fill in a title.

showTitle: Set to false for no title.

xLabelText: Can choose an x axis label.

yLabelText: Can choose a y axis label.

labelSize: Size of x and y axis labels.

Notes

Returns a tuple containing:

model: the PCA model.

metadata: the metadata of individuals included in the PCA.

values: PC values of individuals (before any flipping of axes).

PC1: PC1 values of individuals(after any flipping).

PC2: PC2 values of individuals (after any flipping).

PCAfig: The PCA figure.

GenomicDiversity.plotPCA — Method

plotPCA(myGenoData::GenoData, groups_to_plot_PCA, group_colors_PCA; sampleSet = "", regionText="", flip1 = false, flip2 = false)

Apply the plotPCA function to a GenoData` object.

GenomicDiversity.standardizeIndHet — Method

standardizeIndHet(windowedIndHet)

Divide each individual's windowed heterozgyosity by the average heterozygosity of all individuals (at each window).

​# Arguments

windowedIndHet: The matrix of individual windowed heterozygosity values (rows are individuals; columns represent windowed genomic regions)

Notes

Returns a matrix of standardized individual windowed heterozygosities.